+++ MiniApp/Classes/JavascriptBridge/MiniAppJavaScriptEnumerations.swift
@@ -12,7 +12,8 @@ enum JavaScriptExecResult: String {
 }
 
 enum MiniAppSupportedSchemes: String {
-    case tel
+    case tel    // used for phone calls
+    case about  // used to reveal internal state and built-in functions (e.g.: alert dialog)
 }
 
 /// List of Device Permissions supported by the SDK that can be requested by a Mini app